requirements.txt
paho-mqtt==2.1.0
pip==24.3.1
setuptools==75.6.0
wheel==0.45.1
